backend/requirements.txt
# Core dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# Firebase
firebase-admin==6.4.0

# Data validation
pydantic==2.5.3
marshmallow==3.20.1

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Development
pylint==3.0.3
black==23.12.1
mypy==1.8.0
types-Flask==1.1.6
types-pytz==2023.3.1.1